factor shared code to validate addon install config into parseAddonInstallConfig
